+++ FT_Libft/src/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: jmeirele <jmeirele@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/17 10:45:41 by jmeirele          #+#    #+#             */
-/*   Updated: 2024/11/21 12:05:47 by jmeirele         ###   ########.fr       */
+/*   Updated: 2024/12/17 14:30:35 by jmeirele         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/libft.h"
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i])
